% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordFrequency.R
\name{word_usage}
\alias{word_usage}
\title{Plots the usage of specific words across the three debates.}
\usage{
word_usage(words)
}
\arguments{
\item{words}{A character vector of words from the 'word_percentage_data' data set.}
}
\description{
Generates a facet wrap of plots, one for each of the words supplied by the parameter vector.
Each plot shows the subject word's usage for each turn of the Presidential debates as a % of all words spoken within that turn.
Regression lines and confidence intervals are also provided along with vertical dotted lines that mark the start of each indvidual debate.
To help create parameter vectors the full list of accepted words is supplied as 'word_percentage_data' and comes with this package.
}
\examples{
words <- c("war", "china", "economy")
word_usage(words)

}
\author{
Paul Hazell \email{paul.hazell@postgrad.plymouth.ac.uk}
}
